refactor(dashboard): interfaces
