+++ src/modules/dashboard/Board/components/CardBoard/services/fetchGetBoardsService.ts
@@ -1,5 +1,5 @@
 import api from "shared/services/api";
-import { IFetchGetBoardsResponse } from "modules/dashboard/Board/components/CardBoard/service/types";
+import { IFetchGetBoardsResponse } from "modules/dashboard/Board/components/CardBoard/services/types";
 
 export const fetchGetBoardsService = async () => {
 	const { data } = await api.get<IFetchGetBoardsResponse[]>("/board");
